Add ability to work in other auth contexts
Add methods and context managers to facilitate working in different
projects or different auth contexts, but sharing the Session and other
config information from the existing cloud. Sharing the Session allows
for version discovery cache to be shared, and keeps the user from
needing to do a bunch of work to get a cloud that's like the current one
but slightly different.

Change-Id: I9190ee725cad00db4e0b6ed8321e3021a703a2e4