Shift voting flag and test_matrix_branch for ansible-devel job
We want the job to vote on the github side, but  not here. That means we
should list it as non-voting in our project pipeline definition and not
in the job definition itself.

We also want the job to run using master of devstack. Override the
test_matrix_branch variable.

Also, fix a comment because it's driving slaweq and I crazy.

Change-Id: Ic9d3983de641dbe618c65b2cbf2dcfa3686575df