Use openstacksdk for most transitive depends
This should get openstacksdk on to people's machines in advance of
porting ansible modules to use sdk directly. It also lets us just use
sdk for all of the transitive depends.

Leave os-client-config as openstacksdk doesn't depend on it and
transitioning to openstack.config is a better task for the rocky cycle.

Change-Id: I6d26785b3aa1fc75fb0ad2d45684d93e9628f61b