remove redundant information from release notes build
The release notes build was configured with two pages that scan the
master branch. Now that reno adds consistent anchors to sections, that
means we have two pages with the same anchors and sphinx will complain.
This patch removes the "mainline" page and changes the title of the
"unreleased" page to show the mainline title. We need to keep the page
called "unreleased" so that the release automation that creates the
stable branches can figure out where to insert the next series-specific
page in the sequence in index.rst.

Change-Id: Ie6ed41a8fadb73a2b9b78426499a41ed30bfcd4d
Signed-off-by: Doug Hellmann <doug@doughellmann.com>