Switch to iterable version of server listing (and expose iterable method)
This adds a new server listing mechanism that will yield back each server
block that is returned; it is useful for large clouds with projects with
a large number of active or deleted instances.

For example, in the current list servers it builds up a few lists which can grow
quite big (and consume a lot of memory) making it hard to list projects servers
that have existed for a long time, but with an iterable verion those lists do
not grow (since they just consume the current chunk).

Also the iterable version is more easily used in programs with progress bars
or anything like that that can take advantage of the knowledge that the listing
is at a given chunk (of many).

Change-Id: I84c15e664fb2f72f563618483f9df1c99d215faf