Change 'Member' role reference to 'member'
Keystone has changed the default role during bootstrap from 'Member'
to 'member' therefore the tests which operated under the assumption
of having the 'Member' role assigned are no longer passing.

This patch addresses this issue, however, long term, it's probably
better to create a role and assign it during the test rather than
assuming one already exists, but that's probably for another patch.

Change-Id: I6e4b7665ab046661dd763a1ff8c8e54e6db2dc18