Add support for static routes
The networking API v2 specification, which is implemented by OpenStack
Neutron, features an optional routes parameter - when updating a router
PUT requests). Static routes are crucial for routers to handle traffic
from subnets not directly connected to a router.  This patch adds the
routes parameter to the OpenStackCloud.update_router method as a list of
dictionaries with destination and nexthop parameters.

Example:
[
   {
      "destination": "179.24.1.0/24",
      "nexthop": "172.24.3.99"
    }
]

Change-Id: I14205b6bb071d0b46967f29b6287f74d8796add8