Remove redundant target in README
An 'example' anchor was being explicitly added that was
redundant with the 'Example' title it preceeded. This
causes failures with `python setup.py check -s -r` which
was added to release validation to prevent pypi package
upload failures.

Since the README is imported into the built docs, and the
docs link to the examples using this hidden anchor, this
renames it to avoid a conflict.

Change-Id: I6947e0ae420402cc9866e3c374626392b1d5194c