Decouple OpenStackCloud from Connection
Revert the openstacksdk subclassing from shade. The idea was to
reduce the workload, but trying to make sure that the Cloud abstraction
in openstacksdk doesn't break shade's contract while we update things is
a ton of work to meet the contract that's not really valuable to people.

Instead, we'll put shade on lifesupport and only accept bugfix patches.

Revert "Make OpenStackCloud a subclass of Connection"

This reverts commit ab3f400064f93391bfe444fe64f184be06c7dbe6.

Revert "Use openstack.config directly for config"

This reverts commit 2b48637b6790139c363f0c62977bfa585a2bc1d9.

Revert "Remove the task manager"

This reverts commit 28e95889a0fd8105d12c20575fa66637564a6f89.

Change-Id: I3f5b5fb26af2f6c0bbaade24a04c3d1f274c8cce