Split parser creation and parser for inventory
Access to the object separate from parsing allows us to use
argparse-manpage to generate a man page in the Debian package
automatically. We also need to set 'prog' explicitly rather
than let it be picked up via argv[0] so that when we load
the parser via argparse-manpage it produces the right value.

Change-Id: I654b4408444f804f900951333a6ebc3372d5037e